trustvote/biometrics/requirements.txt
flask==2.3.3
face-recognition==1.3.0
numpy==1.24.3
opencv-python==4.8.0.76
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0 